feat: server.jsをTypeScriptに変更した。
