--- src/utils/.ts
@@ -1,42 +1,0 @@
-fetch("https://www.pixiv.net/ajax/search/illustrations/%E5%B1%B1%E9%A2%A8%20R-18?word=%E5%B1%B1%E9%A2%A8%20R-18&order=date_d&mode=&p=1&csw=0&s_mode=s_tag&type=illust_and_ugoira&lang=ja", {
-    "headers": {
-      "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
-      "accept-language": "ja-JP,ja;q=0.9,en-US;q=0.8,en;q=0.7",
-      "cache-control": "max-age=0",
-      "priority": "u=0, i",
-      "sec-ch-ua": "\"Not/A)Brand\";v=\"8\", \"Chromium\";v=\"126\", \"Google Chrome\";v=\"126\"",
-      "sec-ch-ua-mobile": "?0",
-      "sec-ch-ua-platform": "\"Windows\"",
-      "sec-fetch-dest": "document",
-      "sec-fetch-mode": "navigate",
-      "sec-fetch-site": "cross-site",
-      "sec-fetch-user": "?1",
-      "upgrade-insecure-requests": "1"
-    },
-    "referrerPolicy": "strict-origin-when-cross-origin",
-    "body": null,
-    "method": "GET",
-    "mode": "cors",
-    "credentials": "include"
-  });
-
-  fetch("https://www.pixiv.net/ajax/search/illustrations/%E5%B1%B1%E9%A2%A8%20R-18?word=%E5%B1%B1%E9%A2%A8%20R-18&order=date_d&mode=&p=1&csw=0&s_mode=s_tag&type=illust_and_ugoira&lang=ja", {
-    "headers": {
-      "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
-      "accept-language": "ja-JP,ja;q=0.9,en-US;q=0.8,en;q=0.7",
-      "cache-control": "max-age=0",
-      "priority": "u=0, i",
-      "sec-ch-ua": "\"Not/A)Brand\";v=\"8\", \"Chromium\";v=\"126\", \"Google Chrome\";v=\"126\"",
-      "sec-ch-ua-mobile": "?0",
-      "sec-ch-ua-platform": "\"Windows\"",
-      "sec-fetch-dest": "document",
-      "sec-fetch-mode": "navigate",
-      "sec-fetch-site": "cross-site",
-      "sec-fetch-user": "?1",
-      "upgrade-insecure-requests": "1",
-      "cookie": "p_ab_d_id=1549339903; privacy_policy_notification=0; a_type=0; b_type=1; c_type=104; yuid_b=QoRjlzM; p_ab_id=3; p_ab_id_2=0; privacy_policy_agreement=6; PHPSESSID=15216347_RAsBi2rKH8IhEwvFX7ZhU8lsIh2mGw9n; privacy_policy_agreement=6; cf_clearance=ZReZgf3sFSNRe1cq8em3bgujt8CNNariZEIOsiYP_AA-1722262574-1.0.1.1-mPW5AkAU4xzatXzuDPEe5C5X5KFlWctCjGZv7PyEMZ4MSk0NlX2Ycg6U4d6EckaNEBZtZCY_ExWxMsT9mBIpbw; _im_vid=01J3ZEPKJB35VZAQ0GVV14M3FH; cto_optout=1; default_service_is_touch=no; __cf_bm=kMU.myUXj443AkPWaaYpBOz3r.WQ1pKTLFSwK0LhVzs-1722341349-1.0.1.1-tYG2yA5CfTGCJDHpTpmvasvCeVOVS2cWcF.9fSdHqL1R4IOKZ_2fGuE0oRO3K_YmyZ5Q0fPWIQNVyqp2jInmRHKoXBct3rfKvtFy_rfJJP8"
-    },
-    "referrerPolicy": "strict-origin-when-cross-origin",
-    "body": null,
-    "method": "GET"
-  });
